Implement classification migration as repair step
